requirements.txt
pytesseract==0.3.10
Pillow==10.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
opencv-python==4.8.1.78
pdf2image==1.16.3
